Time: 6 ms (2.23%), Space: 16.1 MB (23.58%) - LeetHub

+++ 0107-binary-tree-level-order-traversal-ii/0107-binary-tree-level-order-traversal-ii.cpp
@@ -37,7 +37,7 @@ public:
             ans.push_back(level);
         }
         vector<vector<int>>res;
-        for(int i=(int)ans.size()-1;i>=0;i--){
+        for(int i=ans.size()-1;i>=0;i--){
             res.push_back(ans[i]);
         }
         return res;
